+++ src/layers/immutable_index.rs
@@ -18,7 +18,6 @@ use std::mem;
 
 use async_trait::async_trait;
 
-use crate::object::*;
 use crate::ops::*;
 use crate::raw::*;
 use crate::*;
